+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/configuration/Locations.java
@@ -125,6 +125,7 @@ public class Locations {
             case BUILDER_CONFIGURATOR:
             case SQL_INSERT_CONFIGURATOR:
             case PROPERTY_ORDER_CONFIGURATOR:
+            case RELATION0_CONFIGURATOR:
             case OUTPUTS_CONFIGURATOR:
             case INPUTS_CONFIGURATOR:
             case SQL_CONFIGURATOR: